requirements.txt
apache-airflow[google]==2.2.5
